refactor pages/play/index.tsx
